chore: move tests to more appropriate file
